Adding new "town" class [WIP]
We have been discussing the fact that real surveys are done by sampling multiple
clusters within a population. There is some vagueness in terminology here; we
previously generated and sampled what we called a “population”, which was a
single spatially contiguous (but not necessarily homogeneous) unit. Now we want
to generate and sample multiple non-contiguous places, which we will call towns
(although the more standard survey terminology would be to call the generated
units, or the samples taken from them, “clusters”); a collection of towns will be
called a population. (Our previous simulations now represent populations that
are made up of a single town.)