Adding new "skip" parameter to EPI sampling method
It is now possible to set how many of the nearest households to skip when selecting the next household.
Setting this parameter to 1 will mean that the nearest household to the current household will be
included in the sample.  Any other number greater than 1 will skip to the Nth nearest household
from the current household.