Fixing iteration-based bugs in EPI samplers and changing "retry" iterations from 100 to 1000
EPI samplers were using the same angle when they were unable to find any buildings in their current
strip/arc.  This caused the sampler to re-try the same strip/arc over and over again, each time not
finding any buildings, and giving up claiming that it couldn't find any buildings in the population.

To correct this for the arc method the angle will increase by the size of the arc width for each
iteration.  For the strip method the angle is increased 2*PI/1000 radians (this so that, after its
1000 attempts, it will eventually cover the entire population).  Both fixes should guarantee that
a building is found.